🚚 Rename: api-constants 대신 url-constants로 파일명 변경 / 팀별 베이스 도메인 추가 (#206)
