--- src/constants/api-constants.ts
@@ -1,2 +1,0 @@
-//TODO: 실제 서버 도메인으로 변경
-export const API_BASE_URL = 'http://example.com'
