+++ src/constants/url-constants.ts
@@ -1,0 +1,9 @@
+export const BASE_URL = 'ozcoding.site'
+
+export const API_BASE_URL = `https://api.${BASE_URL}/api/v1`
+
+// 1팀 담당 - 스터디 그룹 베이스 도메인
+export const STUDYGROUP_BASE_URL = `https://study.${BASE_URL}/study-group`
+
+// 3팀 담당 - 채용 + 강의 베이스 도메인
+export const LEARN_BASE_URL = `https://learn.${BASE_URL}/recruitment`
