+++ src/mocks/handlers/bookmarked-recruitments-handler.ts
@@ -1,4 +1,4 @@
-import { API_BASE_URL } from '@/constants/api-constants'
+import { API_BASE_URL } from '@/constants/url-constants'
 import { http, HttpResponse } from 'msw'
 import { bookmarkedRecruitmentsMock } from '@/mocks/data/bookmarked-recruitments-data'
