+++ src/mocks/handlers/chat-handler.ts
@@ -1,4 +1,4 @@
-import { API_BASE_URL } from '@/constants/api-constants'
+import { API_BASE_URL } from '@/constants/url-constants'
 import { http, HttpResponse, ws } from 'msw'
 import { chatRoomMessages } from '@/mocks/data/chat-room-data'
