+++ src/mocks/handlers/notification-handler.ts
@@ -1,6 +1,6 @@
 import { http, HttpResponse } from 'msw'
 import { mockNotificationList } from '@/mocks/data/notification-data'
-import { API_BASE_URL } from '@/constants/api-constants'
+import { API_BASE_URL } from '@/constants/url-constants'
 
 const getNotifications = http.get(`${API_BASE_URL}/notifications`, () => {
   return HttpResponse.json(mockNotificationList)
