+++ src/utils/axios.ts
@@ -1,4 +1,4 @@
-import { API_BASE_URL } from '@/constants/api-constants'
+import { API_BASE_URL } from '@/constants/url-constants'
 import axios from 'axios'
 
 const api = axios.create({
